+++ fractals/src/main/java/me/catmousedog/fractals/canvas/Field.java
@@ -1,43 +1,64 @@
 package me.catmousedog.fractals.canvas;
 
 import java.awt.image.BufferedImage;
-import java.util.ArrayList;
-import java.util.List;
 
 import org.jetbrains.annotations.NotNull;
 
-import me.catmousedog.fractals.fractals.Pixel;
-
 public class Field {
-	
+
 	private BufferedImage img;
-	
-	private List<Pixel> field;
-	
+
+	private Pixel[] pixels;
+
 	public Field(int width, int height) {
 		setSize(width, height);
 	}
-	
+
 	public void setSize(int width, int height) {
 		img = new BufferedImage(width, height, BufferedImage.TYPE_INT_RGB);
-		
+
 		// set field size
-		field = new ArrayList<Pixel>(width * height);
+		pixels = new Pixel[width * height];
+		int i = 0;
 		for (int x = 0; x < width; x++) {
 			for (int y = 0; y < height; y++) {
-				field.add(new Pixel(x, y));
+				pixels[i] = new Pixel(x, y);
+				i++;
 			}
 		}
 	}
-	
+
+	/**
+	 * Finds the <code>Pixel</code> with the same <code>x</code> and <code>y</code>
+	 * coordinates.<br>
+	 * This is done by comparing the hash code of each <code>Pixel</code> with the
+	 * <code>Pixel</code> at <code>x:y</code>.
+	 * <p>
+	 * The comparing of hash codes is effective up to an <code>x</code> and
+	 * <code>y</code> of 2^16 or 65536.
+	 * 
+	 * @param x the actual <code>x</code> coordinate
+	 * @param y the actual <code>y</code> coordinate
+	 * @return the <code>Pixel</code> with the same <code>x</code> and
+	 *         <code>y</code> coordinates.
+	 */
+	public Pixel getPixel(int x, int y) {
+		int h = new Pixel(x, y).hashCode();
+		for (Pixel p : pixels) {
+			if (p.hashCode() == h)
+				return p;
+		}
+		return null;
+	}
+
 	@NotNull
 	public BufferedImage getImg() {
 		return img;
 	}
-	
+
 	@NotNull
-	public List<Pixel> getPixels() {
-		return field;
+	public Pixel[] getPixels() {
+		return pixels;
 	}
-	
+
 }