+++ fractals/src/main/java/me/catmousedog/fractals/canvas/Mouse.java
@@ -1,8 +1,13 @@
 package me.catmousedog.fractals.canvas;
 
+import java.awt.Point;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
+import java.awt.event.MouseMotionListener;
 
+import javax.swing.JLabel;
+import javax.swing.Popup;
+import javax.swing.PopupFactory;
 import javax.swing.SwingUtilities;
 
 import me.catmousedog.fractals.fractals.LinearTransform;
@@ -11,7 +16,7 @@ import me.catmousedog.fractals.ui.JPInterface;
 /**
  * the mouse listener for interacting with the canvas
  */
-public class Mouse implements MouseListener {
+public class Mouse implements MouseListener, MouseMotionListener {
 
 	/**
 	 * the instance of the canvas this MouseListener belongs to
@@ -35,41 +40,91 @@ public class Mouse implements MouseListener {
 		// JPInterface#renderWithout(Runnable r) couldn't be used because of the
 		// MouseEvent here which isn't effectively final
 
-		jpi.save();
-
-		LinearTransform transform = canvas.getFractal().getTransform();
-		double[] t = transform.apply(me.getX(), me.getY());
-		transform.setTranslation(t[0], t[1]);
-
 		// lmb (zoom in)
-		if (me.getButton() == MouseEvent.BUTTON1)
+		if (me.getButton() == MouseEvent.BUTTON1) {
+			jpi.save();
+			LinearTransform transform = canvas.getFractal().getTransform();
+			double[] t = transform.apply(me.getX(), me.getY());
+			transform.setTranslation(t[0], t[1]);
 			transform.zoom(1 / canvas.getZoomFactor());
 
+			jpi.update();
+			jpi.renderNow();
+		}
+
 		// rmb (zoom out)
-		if (me.getButton() == MouseEvent.BUTTON3)
+		if (me.getButton() == MouseEvent.BUTTON3) {
+			jpi.save();
+			LinearTransform transform = canvas.getFractal().getTransform();
+			double[] t = transform.apply(me.getX(), me.getY());
+			transform.setTranslation(t[0], t[1]);
 			transform.zoom(canvas.getZoomFactor());
 
-		jpi.update();
-		jpi.renderNow();
+			jpi.update();
+			jpi.renderNow();
+		}
+
+		// mmb (info)
+		if (me.getButton() == MouseEvent.BUTTON2) {
+			middleMouse = false;
+			tip.hide();
+		}
+
 	}
 
 	@Override
-	public void mouseClicked(MouseEvent e) {
+	public void mousePressed(MouseEvent me) {
+		me = SwingUtilities.convertMouseEvent(me.getComponent(), me, canvas);
+
+		// mmb (info)
+		if (me.getButton() == MouseEvent.BUTTON2) {
+			middleMouse = true;
+			displayTip(me.getX(), me.getY());
+		}
 	}
 
 	@Override
-	public void mousePressed(MouseEvent e) {
+	public void mouseDragged(MouseEvent me) {
+		if (!middleMouse)
+			return;
+
+		me = SwingUtilities.convertMouseEvent(me.getComponent(), me, canvas);
+
+		tip.hide();
+		displayTip(me.getX(), me.getY());
 	}
 
 	@Override
-	public void mouseEntered(MouseEvent e) {
+	public void mouseClicked(MouseEvent me) {
 	}
 
 	@Override
-	public void mouseExited(MouseEvent e) {
+	public void mouseEntered(MouseEvent me) {
+	}
+
+	@Override
+	public void mouseExited(MouseEvent me) {
+	}
+
+	@Override
+	public void mouseMoved(MouseEvent me) {
+	}
+
+	private Popup tip;
+
+	private boolean middleMouse = false;
+
+	private void displayTip(int x, int y) {
+		PopupFactory popupFactory = PopupFactory.getSharedInstance();
+		Point screenLocation = canvas.getLocationOnScreen();
+		Point location = new Point(screenLocation.x + x, screenLocation.y + y - 20);
+		Pixel pixel = canvas.getField().getPixel(x, y);
+		tip = popupFactory.getPopup(canvas, new JLabel(pixel.toString()), location.x, location.y);
+		tip.show();
 	}
 
 	public void setJPI(JPInterface jpi) {
 		this.jpi = jpi;
 	}
+
 }