+++ CHANGELOG.md
@@ -217,9 +217,18 @@ The original idea of making the **JPInterface** a class to access and retrieve d
 - *Picture* section
 - *Settings.addImage()*
 - TextField min and max for safety input
+- ComboBox for predefined resolutions
+- concrete locations
+- middle mouse button shows Pixel information
 
 ### Changed
 - Moved **AllData** to new **GUI** class and new source file
+- Moved 'settings' code from **Fractal** to **Settings**
+- List of Pixels to Pixel array. Tremendously increase the parallel speed! colouring: before=300ms after=40ms<br> This also solved the Pixel creating issue when trying to create huge images. The lag spike is now 2 seconds compared to 20 seconds.
+- locations appear in order in combobox
 
 ### Removed
-- Listener from TextField
+- Listener from TextFields
+
+### Note
+- Did some testing on the efficiency of HashSet and HashMaps for storing the Pixels, the array seems to be the fasted option for parallel execution